File restructure for release
