+++ experimental/piranha_playground/data_validation.py
@@ -11,8 +11,8 @@
 
 import attr
 import toml
-from experimental.rule_inference.utils.pretty_toml import PrettyTOML
-from experimental.rule_inference.utils.rule_utils import RawRuleGraph
+from piranha_playground.rule_inference.utils.pretty_toml import PrettyTOML
+from piranha_playground.rule_inference.utils.rule_utils import RawRuleGraph
 
 
 @attr.s
