+++ experimental/piranha_playground/rule_inference/graph_parser.py
@@ -13,7 +13,7 @@ from collections import defaultdict, deque
 from typing import Deque, Dict, List, Set, Tuple
 
 import attr
-from experimental.rule_inference.utils.node_utils import NodeUtils
+from piranha_playground.rule_inference.utils.node_utils import NodeUtils
 from tree_sitter import Node, Tree
