+++ experimental/piranha_playground/rule_inference/piranha_chat.py
@@ -17,7 +17,7 @@ from typing import List, Optional, Tuple
 
 import attr
 import openai
-from experimental.rule_inference.utils.logger_formatter import CustomFormatter
+from piranha_playground.rule_inference.utils.logger_formatter import CustomFormatter
 
 logger = logging.getLogger("PiranhaAgent")
 logger.setLevel(logging.DEBUG)
